feat(rfc-112 stage 1): the store refuses, and writes nothing doing it
One reader and one rule (D4): read_stored_version and check_supported, used by
the runner, backup create, restore, verify, the settings page and the test-only
runner. Verified: the only SELECTs against sui_meta left in the tree are this
reader, last_migrated_by, and state.rs's unrelated key.

The check runs first, on a read-only connection (D3). check_database_file is
the first line of Database::open, before Connection::open and before the
journal-mode pragma -- which is itself a write, and used to convert a
rollback-journal database to WAL before anything could refuse it.

Fresh means no application tables, not no row (D2). A populated database with
no version row, and a foreign SQLite file, are refused instead of migrated; the
foreign file used to gain 26 tables.

The canonical-form check and the i64 parse close D6's hole together: all-digit
values that overflow can only be too new, so they are reported too new rather
than invalid, and +43, leading zeros, spaces, reals, hex and BLOBs are invalid.
An explicit 0 in a populated database is invalid too, which the RFC did not
name and the implementer found.

run reads and judges inside BEGIN IMMEDIATE before creating sui_meta, and
apply_migration re-reads under its own; the race test failed 60 of 80 opens
against the unfixed runner.

Restore now reads the version of the database inside the archive, so a lying
manifest and a manifest-less archive can no longer carry a too-new or
unreadable database past the check. Backup create records a too-new database
truthfully rather than stamping 0 -- a backup is the operator's way out of a
refusal.

942 to 969 tests, 13 mutations each shown failing the right tests. Verified by
the reviewer: 23 of 23 hunks and both new files match, nothing unclaimed; fmt,
clippy -D warnings, 969 tests and every doc gate re-run here.

One correction by the reviewer: docs/src/guides/upgrade.md said an older binary
"starts without complaint", which this commit makes false. It now states the
refusal and that nothing is written. The operator line, exit 65 and the tracing
event are stage 2, and the guide does not describe them yet.

+++ crates/sui-id-web/src/pages/settings/other.rs
@@ -5,7 +5,9 @@ use crate::layout::Shell;
 
 pub struct SettingsOtherData {
     pub binary_version: String,
-    pub schema_version: i32,
+    /// The schema version **stored in the database** (RFC 112 D4), not the
+    /// binary's ceiling: the two differ exactly when it matters.
+    pub schema_version: i64,
     pub db_path: String,
     pub master_key_file: String,
     pub user_count: usize,
